Cleaned up movement.c
